+++ package.json
@@ -1,10 +1,12 @@
 {
   "name": "@zoralabs/chains",
   "version": "1.3.0",
-  "main": "index.js",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
   "files": [
     "bin",
-    "./index.js"
+    "dist",
+    "*.ts"
   ],
   "bin": {
     "chains": "./bin/chains"
